requirements.txt
google-adk
python-dotenv
yfinance
requests
